@@ -2,7 +2,7 @@
   "private": true,
   "packageManager": "pnpm@7.9.0",
   "engines": {
-    "node": "^14"
+    "node": "^14.19.0"
   },
   "scripts": {
     "build": "nuxi build",
