@@ -1,6 +1,5 @@
 import chromium from 'chrome-aws-lambda'
-
-const playwright = require('playwright-core')
+import playwright from 'playwright-core'
 
 exports.handler = async function (event, context) {
   const browser = await playwright.chromium.launch({
